Refactor: Name modify.
