--- src/main/java/com/sejong/creativesemester/coucil/reposiotory/CouncilRepositoryCustom.java
@@ -1,7 +1,0 @@
-package com.sejong.creativesemester.coucil.reposiotory;
-
-import com.sejong.creativesemester.coucil.reposiotory.res.CouncilInfoResponseDto;
-
-public interface CouncilRepositoryCustom {
-    CouncilInfoResponseDto findCouncilInfoByMajorId(Long majorId);
-}
