+++ src/main/java/com/sejong/creativesemester/council/repository/CouncilRepositoryCustom.java
@@ -1,0 +1,7 @@
+package com.sejong.creativesemester.council.repository;
+
+import com.sejong.creativesemester.council.repository.res.CouncilInfoResponseDto;
+
+public interface CouncilRepositoryCustom {
+    CouncilInfoResponseDto findCouncilInfoByMajorId(Long majorId);
+}
